+++ packages/engine/src/physics/behaviors/physicalPrimitives.ts
@@ -1,4 +1,4 @@
-import { threeToCannon } from '@xr3ngine/engine/src/templates/world/three-to-cannon';
+import { threeToCannon } from '../classes/three-to-cannon';
 import { Body, Box, ConvexPolyhedron, Cylinder, Plane, Quaternion, Sphere, Vec3 } from 'cannon-es';
 import { Entity } from '../../ecs/classes/Entity';
 import { getComponent } from '../../ecs/functions/EntityFunctions';
