+++ packages/engine/src/templates/character/CharacterAvatars.ts
@@ -25,7 +25,6 @@ export const defaultAvatarAnimations: {[key:number]: AnimationConfigInterface} =
   [CharacterStateTypes.IDLE_ROTATE_RIGHT]: { name: 'turn_right' },
   [CharacterStateTypes.IDLE_ROTATE_LEFT]: { name: 'left_turn' },
   [CharacterStateTypes.JUMP]: { name: 'jump_up', loop: LoopOnce },
-  [CharacterStateTypes.JUMP_RUNNING]: { name: 'jump', loop: LoopOnce },
   [CharacterStateTypes.FALLING]: { name: 'falling' },
   [CharacterStateTypes.FALLING_LONG]: { name: 'falling' },
   [CharacterStateTypes.DROP]: { name: 'falling_to_land' },
