+++ packages/engine/src/templates/interactive/interfaces/CommonInteractiveData.ts
@@ -1,5 +1,4 @@
-import { CommonInteractiveAction } from "../types/CommonInteractiveAction";
-
+type CommonInteractiveAction = "link" | "infoBox" | "mediaSource";
 export interface CommonInteractiveData {
   action: CommonInteractiveAction;
   payload: CommonInteractiveDataPayload;
