--- packages/engine/src/templates/interactive/types/CommonInteractiveAction.ts
@@ -1,1 +1,0 @@
-export type CommonInteractiveAction = "link" | "infoBox" | "mediaSource";
