--- packages/engine/src/templates/networking/DefaultMessageTypes.ts
@@ -1,3 +1,0 @@
-export enum DefaultMessageTypes {
-
-}
