+++ packages/engine/src/templates/networking/DefaultNetworkSchema.ts
@@ -1,9 +1,24 @@
 import { MessageTypes } from '../../networking/enums/MessageTypes';
 import { NetworkSchema } from '../../networking/interfaces/NetworkSchema';
 import { NetworkPlayerCharacter } from '../character/prefabs/NetworkPlayerCharacter';
-import { NetworkWorldObject } from '../world/prefabs/NetworkWorldObject';
 import { NetworkVehicle } from '../car/prefabs/NetworkVehicle';
-import { DefaultMessageTypes } from './DefaultMessageTypes';
+import { NetworkPrefab } from '../../networking/interfaces/NetworkPrefab';
+import { TransformComponent } from '../../transform/components/TransformComponent';
+
+// Prefab is a pattern for creating an entity and component collection as a prototype
+export const NetworkWorldObject: NetworkPrefab = {
+  // These will be created for all players on the network
+  networkComponents: [
+    // Transform system applies values from transform component to three.js object (position, rotation, etc)
+    { type: TransformComponent }
+    // Local player input mapped to behaviors in the input map
+  ],
+  // These are only created for the local player who owns this prefab
+  localClientComponents: [],
+  serverComponents: [],
+  onAfterCreate: [],
+  onBeforeDestroy: []
+};
 
 export const PrefabType = {
   Player: 0,
@@ -20,8 +35,7 @@ export const DefaultPrefabs = {
 export const DefaultNetworkSchema: NetworkSchema = {
   transport: null,
   messageTypes: {
-    ...MessageTypes,
-    ...DefaultMessageTypes
+    ...MessageTypes
   },
   prefabs: DefaultPrefabs,
   defaultClientPrefab: PrefabType.Player
