+++ packages/engine/src/templates/vehicle/VehicleInputSchema.ts
