+++ packages/engine/src/templates/vehicle/behaviors/playerModelInCar.ts
