+++ packages/engine/src/templates/vehicle/prefabs/NetworkVehicle.ts
